refactor: 디테일 페이지 컴포넌트 파일명 카멜케이스 적용
