+++ src/components/features/home/background-handler.tsx
